@@ -1,78 +1,27 @@
 {
-  "_from": "react-router-dom",
-  "_id": "react-router-dom@5.2.0",
-  "_inBundle": false,
-  "_integrity": "sha512-gxAmfylo2QUjcwxI63RhQ5G85Qqt4voZpUXSEqCwykV0baaOTQDR1f0PmY8AELqIyVc0NEZUj0Gov5lNGcXgsA==",
-  "_location": "/react-router-dom",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "tag",
-    "registry": true,
-    "raw": "react-router-dom",
-    "name": "react-router-dom",
-    "escapedName": "react-router-dom",
-    "rawSpec": "",
-    "saveSpec": null,
-    "fetchSpec": "latest"
+  "name": "react-router-dom",
+  "version": "6.2.1",
+  "author": "Remix Software <hello@remix.run>",
+  "description": "Declarative routing for React web applications",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/remix-run/react-router.git",
+    "directory": "packages/react-router-dom"
   },
-  "_requiredBy": [
-    "#USER",
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/react-router-dom/-/react-router-dom-5.2.0.tgz",
-  "_shasum": "9e65a4d0c45e13289e66c7b17c7e175d0ea15662",
-  "_spec": "react-router-dom",
-  "_where": "/Users/kabash/Documents/shop-frontend",
-  "author": {
-    "name": "React Training",
-    "email": "hello@reacttraining.com"
-  },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
-  },
-  "bugs": {
-    "url": "https://github.com/ReactTraining/react-router/issues"
-  },
-  "bundleDependencies": false,
+  "license": "MIT",
+  "main": "./main.js",
+  "module": "./index.js",
+  "types": "./index.d.ts",
+  "unpkg": "./umd/react-router-dom.production.min.js",
   "dependencies": {
-    "@babel/runtime": "^7.1.2",
-    "history": "^4.9.0",
-    "loose-envify": "^1.3.1",
-    "prop-types": "^15.6.2",
-    "react-router": "5.2.0",
-    "tiny-invariant": "^1.0.2",
-    "tiny-warning": "^1.0.0"
+    "react-router": "6.2.1",
+    "history": "^5.2.0"
   },
-  "deprecated": false,
-  "description": "DOM bindings for React Router",
-  "files": [
-    "BrowserRouter.js",
-    "HashRouter.js",
-    "Link.js",
-    "MemoryRouter.js",
-    "NavLink.js",
-    "Prompt.js",
-    "Redirect.js",
-    "Route.js",
-    "Router.js",
-    "StaticRouter.js",
-    "Switch.js",
-    "cjs",
-    "es",
-    "esm",
-    "index.js",
-    "generatePath.js",
-    "matchPath.js",
-    "modules/*.js",
-    "modules/utils/*.js",
-    "withRouter.js",
-    "warnAboutDeprecatedCJSRequire.js",
-    "umd"
-  ],
-  "gitHead": "21a62e55c0d6196002bd4ab5b3350514976928cf",
-  "homepage": "https://github.com/ReactTraining/react-router#readme",
+  "peerDependencies": {
+    "react": ">=16.8",
+    "react-dom": ">=16.8"
+  },
+  "sideEffects": false,
   "keywords": [
     "react",
     "router",
@@ -80,22 +29,5 @@
     "routing",
     "history",
     "link"
-  ],
-  "license": "MIT",
-  "main": "index.js",
-  "module": "esm/react-router-dom.js",
-  "name": "react-router-dom",
-  "peerDependencies": {
-    "react": ">=15"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ReactTraining/react-router.git"
-  },
-  "scripts": {
-    "build": "rollup -c",
-    "lint": "eslint modules"
-  },
-  "sideEffects": false,
-  "version": "5.2.0"
+  ]
 }
