@@ -1,77 +1,25 @@
 {
-  "_from": "react-router",
-  "_id": "react-router@5.2.0",
-  "_inBundle": false,
-  "_integrity": "sha512-smz1DUuFHRKdcJC0jobGo8cVbhO3x50tCL4icacOlcwDOEQPq4TMqwx3sY1TP+DvtTgz4nm3thuo7A+BK2U0Dw==",
-  "_location": "/react-router",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "tag",
-    "registry": true,
-    "raw": "react-router",
-    "name": "react-router",
-    "escapedName": "react-router",
-    "rawSpec": "",
-    "saveSpec": null,
-    "fetchSpec": "latest"
-  },
-  "_requiredBy": [
-    "#USER",
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/react-router/-/react-router-5.2.0.tgz",
-  "_shasum": "424e75641ca8747fbf76e5ecca69781aa37ea293",
-  "_spec": "react-router",
-  "_where": "/Users/kabash/Documents/shop-frontend",
-  "author": {
-    "name": "React Training",
-    "email": "hello@reacttraining.com"
-  },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
-  },
-  "bugs": {
-    "url": "https://github.com/ReactTraining/react-router/issues"
-  },
-  "bundleDependencies": false,
-  "dependencies": {
-    "@babel/runtime": "^7.1.2",
-    "history": "^4.9.0",
-    "hoist-non-react-statics": "^3.1.0",
-    "loose-envify": "^1.3.1",
-    "mini-create-react-context": "^0.4.0",
-    "path-to-regexp": "^1.7.0",
-    "prop-types": "^15.6.2",
-    "react-is": "^16.6.0",
-    "tiny-invariant": "^1.0.2",
-    "tiny-warning": "^1.0.0"
-  },
-  "deprecated": false,
+  "name": "react-router",
+  "version": "6.2.1",
+  "author": "Remix Software <hello@remix.run>",
   "description": "Declarative routing for React",
-  "files": [
-    "MemoryRouter.js",
-    "Prompt.js",
-    "Redirect.js",
-    "Route.js",
-    "Router.js",
-    "StaticRouter.js",
-    "Switch.js",
-    "cjs",
-    "es",
-    "esm",
-    "index.js",
-    "generatePath.js",
-    "matchPath.js",
-    "modules/*.js",
-    "modules/utils/*.js",
-    "withRouter.js",
-    "warnAboutDeprecatedCJSRequire.js",
-    "umd"
-  ],
-  "gitHead": "21a62e55c0d6196002bd4ab5b3350514976928cf",
-  "homepage": "https://github.com/ReactTraining/react-router#readme",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/remix-run/react-router.git",
+    "directory": "packages/react-router"
+  },
+  "license": "MIT",
+  "main": "./main.js",
+  "module": "./index.js",
+  "types": "./index.d.ts",
+  "unpkg": "./umd/react-router.production.min.js",
+  "peerDependencies": {
+    "react": ">=16.8"
+  },
+  "dependencies": {
+    "history": "^5.2.0"
+  },
+  "sideEffects": false,
   "keywords": [
     "react",
     "router",
@@ -79,22 +27,5 @@
     "routing",
     "history",
     "link"
-  ],
-  "license": "MIT",
-  "main": "index.js",
-  "module": "esm/react-router.js",
-  "name": "react-router",
-  "peerDependencies": {
-    "react": ">=15"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ReactTraining/react-router.git"
-  },
-  "scripts": {
-    "build": "rollup -c",
-    "lint": "eslint modules"
-  },
-  "sideEffects": false,
-  "version": "5.2.0"
+  ]
 }
